refactor(seller): seller-product-crud 4분할 (P0-2 단계 1)
SellerProductCrudService (720줄/11 public) 를 책임별 4개 서비스로 분리.
각 서비스는 interface + Symbol DI Token 패턴 (P0-1과 일관).

- 신규: SellerProductQueryService (sellerProducts, sellerProduct)
- 신규: SellerProductLifecycleService (create/update/delete/setActive)
  - private: validateProductPrices, buildProductUpdateData
- 신규: SellerProductImageService (add/delete/reorder image)
- 신규: SellerProductTaxonomyService (setCategories/setTags)
- 신규 헬퍼: seller-product-mappers.helper.ts (toProductOutput,
  toOptionGroupOutput, toCustomTemplateOutput, toProductImageOutput
  + ProductDetailRow 등 row 타입) — 4개 서비스가 공유
- 제거: SellerProductCrudService + 해당 spec
- Resolver 갱신:
  - SellerProductQueryResolver → SELLER_PRODUCT_QUERY_SERVICE 주입
  - SellerProductMutationResolver → 3 신규 service 주입 (lifecycle/image/taxonomy)
- SellerModule: 4 신규 service 등록 (SELLER_PRODUCT_*_SERVICE 토큰)
- Spec 분할: seller-product-crud.service.spec → 4 spec (query/lifecycle/image/taxonomy)
- resolver spec: 4 신규 service provider 등록

각 서비스는 SellerBaseService 를 extends (requireSellerContext / cleanCurrency 등 공유).
FE 영향 없음 (Resolver method 시그니처 동일, SDL 동일).
