+++ src/features/seller/resolvers/seller-product-mutation.resolver.ts
@@ -1,4 +1,4 @@
-import { UseGuards } from '@nestjs/common';
+import { Inject, UseGuards } from '@nestjs/common';
 import { Args, Mutation, Resolver } from '@nestjs/graphql';
 
 import { parseId } from '@/common/utils/id-parser';
@@ -21,7 +21,18 @@ import { SellerUpsertProductCustomTemplateInput } from '@/features/seller/dto/in
 import { SellerUpsertProductCustomTextTokenInput } from '@/features/seller/dto/inputs/seller-upsert-product-custom-text-token.input';
 import { SellerCustomTemplateService } from '@/features/seller/services/seller-custom-template.service';
 import { SellerOptionService } from '@/features/seller/services/seller-option.service';
-import { SellerProductCrudService } from '@/features/seller/services/seller-product-crud.service';
+import {
+  SELLER_PRODUCT_IMAGE_SERVICE,
+  type ISellerProductImageService,
+} from '@/features/seller/services/seller-product-image.service.interface';
+import {
+  SELLER_PRODUCT_LIFECYCLE_SERVICE,
+  type ISellerProductLifecycleService,
+} from '@/features/seller/services/seller-product-lifecycle.service.interface';
+import {
+  SELLER_PRODUCT_TAXONOMY_SERVICE,
+  type ISellerProductTaxonomyService,
+} from '@/features/seller/services/seller-product-taxonomy.service.interface';
 import type {
   SellerCustomTemplateOutput,
   SellerCustomTextTokenOutput,
@@ -41,7 +52,12 @@ import {
 @UseGuards(JwtAuthGuard)
 export class SellerProductMutationResolver {
   constructor(
-    private readonly productService: SellerProductCrudService,
+    @Inject(SELLER_PRODUCT_LIFECYCLE_SERVICE)
+    private readonly productLifecycle: ISellerProductLifecycleService,
+    @Inject(SELLER_PRODUCT_IMAGE_SERVICE)
+    private readonly productImage: ISellerProductImageService,
+    @Inject(SELLER_PRODUCT_TAXONOMY_SERVICE)
+    private readonly productTaxonomy: ISellerProductTaxonomyService,
     private readonly optionService: SellerOptionService,
     private readonly templateService: SellerCustomTemplateService,
   ) {}
@@ -52,7 +68,7 @@ export class SellerProductMutationResolver {
     @Args('input') input: SellerCreateProductInput,
   ): Promise<SellerProductOutput> {
     const accountId = parseAccountId(user);
-    return this.productService.sellerCreateProduct(accountId, input);
+    return this.productLifecycle.sellerCreateProduct(accountId, input);
   }
 
   @Mutation('sellerUpdateProduct')
@@ -61,7 +77,7 @@ export class SellerProductMutationResolver {
     @Args('input') input: SellerUpdateProductInput,
   ): Promise<SellerProductOutput> {
     const accountId = parseAccountId(user);
-    return this.productService.sellerUpdateProduct(accountId, input);
+    return this.productLifecycle.sellerUpdateProduct(accountId, input);
   }
 
   @Mutation('sellerDeleteProduct')
@@ -70,7 +86,7 @@ export class SellerProductMutationResolver {
     @Args('productId') productId: string,
   ): Promise<boolean> {
     const accountId = parseAccountId(user);
-    return this.productService.sellerDeleteProduct(
+    return this.productLifecycle.sellerDeleteProduct(
       accountId,
       parseId(productId),
     );
@@ -82,7 +98,7 @@ export class SellerProductMutationResolver {
     @Args('input') input: SellerSetProductActiveInput,
   ): Promise<SellerProductOutput> {
     const accountId = parseAccountId(user);
-    return this.productService.sellerSetProductActive(accountId, input);
+    return this.productLifecycle.sellerSetProductActive(accountId, input);
   }
 
   @Mutation('sellerAddProductImage')
@@ -91,7 +107,7 @@ export class SellerProductMutationResolver {
     @Args('input') input: SellerAddProductImageInput,
   ): Promise<SellerProductImageOutput> {
     const accountId = parseAccountId(user);
-    return this.productService.sellerAddProductImage(accountId, input);
+    return this.productImage.sellerAddProductImage(accountId, input);
   }
 
   @Mutation('sellerDeleteProductImage')
@@ -100,7 +116,7 @@ export class SellerProductMutationResolver {
     @Args('imageId') imageId: string,
   ): Promise<boolean> {
     const accountId = parseAccountId(user);
-    return this.productService.sellerDeleteProductImage(
+    return this.productImage.sellerDeleteProductImage(
       accountId,
       parseId(imageId),
     );
@@ -112,7 +128,7 @@ export class SellerProductMutationResolver {
     @Args('input') input: SellerReorderProductImagesInput,
   ): Promise<SellerProductImageOutput[]> {
     const accountId = parseAccountId(user);
-    return this.productService.sellerReorderProductImages(accountId, input);
+    return this.productImage.sellerReorderProductImages(accountId, input);
   }
 
   @Mutation('sellerSetProductCategories')
@@ -121,7 +137,7 @@ export class SellerProductMutationResolver {
     @Args('input') input: SellerSetProductCategoriesInput,
   ): Promise<SellerProductOutput> {
     const accountId = parseAccountId(user);
-    return this.productService.sellerSetProductCategories(accountId, input);
+    return this.productTaxonomy.sellerSetProductCategories(accountId, input);
   }
 
   @Mutation('sellerSetProductTags')
@@ -130,7 +146,7 @@ export class SellerProductMutationResolver {
     @Args('input') input: SellerSetProductTagsInput,
   ): Promise<SellerProductOutput> {
     const accountId = parseAccountId(user);
-    return this.productService.sellerSetProductTags(accountId, input);
+    return this.productTaxonomy.sellerSetProductTags(accountId, input);
   }
 
   @Mutation('sellerCreateOptionGroup')
