+++ src/features/seller/resolvers/seller-product-query.resolver.ts
@@ -1,9 +1,12 @@
-import { UseGuards } from '@nestjs/common';
+import { Inject, UseGuards } from '@nestjs/common';
 import { Args, Query, Resolver } from '@nestjs/graphql';
 
 import { parseId } from '@/common/utils/id-parser';
 import { SellerProductListInput } from '@/features/seller/dto/inputs/seller-product-list.input';
-import { SellerProductCrudService } from '@/features/seller/services/seller-product-crud.service';
+import {
+  SELLER_PRODUCT_QUERY_SERVICE,
+  type ISellerProductQueryService,
+} from '@/features/seller/services/seller-product-query.service.interface';
 import type {
   SellerCursorConnection,
   SellerProductOutput,
@@ -18,7 +21,10 @@ import {
 @Resolver('Query')
 @UseGuards(JwtAuthGuard)
 export class SellerProductQueryResolver {
-  constructor(private readonly productService: SellerProductCrudService) {}
+  constructor(
+    @Inject(SELLER_PRODUCT_QUERY_SERVICE)
+    private readonly productQuery: ISellerProductQueryService,
+  ) {}
 
   @Query('sellerProducts')
   sellerProducts(
@@ -26,7 +32,7 @@ export class SellerProductQueryResolver {
     @Args('input', { nullable: true }) input?: SellerProductListInput,
   ): Promise<SellerCursorConnection<SellerProductOutput>> {
     const accountId = parseAccountId(user);
-    return this.productService.sellerProducts(accountId, input);
+    return this.productQuery.sellerProducts(accountId, input);
   }
 
   @Query('sellerProduct')
@@ -35,6 +41,6 @@ export class SellerProductQueryResolver {
     @Args('productId') productId: string,
   ): Promise<SellerProductOutput> {
     const accountId = parseAccountId(user);
-    return this.productService.sellerProduct(accountId, parseId(productId));
+    return this.productQuery.sellerProduct(accountId, parseId(productId));
   }
 }
