+++ src/features/seller/resolvers/seller-product.resolver.spec.ts
@@ -25,7 +25,14 @@ import { SellerProductMutationResolver } from '@/features/seller/resolvers/selle
 import { SellerProductQueryResolver } from '@/features/seller/resolvers/seller-product-query.resolver';
 import { SellerCustomTemplateService } from '@/features/seller/services/seller-custom-template.service';
 import { SellerOptionService } from '@/features/seller/services/seller-option.service';
-import { SellerProductCrudService } from '@/features/seller/services/seller-product-crud.service';
+import { SellerProductImageService } from '@/features/seller/services/seller-product-image.service';
+import { SELLER_PRODUCT_IMAGE_SERVICE } from '@/features/seller/services/seller-product-image.service.interface';
+import { SellerProductLifecycleService } from '@/features/seller/services/seller-product-lifecycle.service';
+import { SELLER_PRODUCT_LIFECYCLE_SERVICE } from '@/features/seller/services/seller-product-lifecycle.service.interface';
+import { SellerProductQueryService } from '@/features/seller/services/seller-product-query.service';
+import { SELLER_PRODUCT_QUERY_SERVICE } from '@/features/seller/services/seller-product-query.service.interface';
+import { SellerProductTaxonomyService } from '@/features/seller/services/seller-product-taxonomy.service';
+import { SELLER_PRODUCT_TAXONOMY_SERVICE } from '@/features/seller/services/seller-product-taxonomy.service.interface';
 import { disconnectTestPrismaClient } from '@/test/db/prisma-test-client';
 import { closeTruncateConnection, truncateAll } from '@/test/db/truncate';
 import { createProduct, setupSellerWithStore } from '@/test/factories';
@@ -41,7 +48,22 @@ describe('Seller Product Resolvers (real DB)', () => {
       providers: [
         SellerProductQueryResolver,
         SellerProductMutationResolver,
-        SellerProductCrudService,
+        {
+          provide: SELLER_PRODUCT_QUERY_SERVICE,
+          useClass: SellerProductQueryService,
+        },
+        {
+          provide: SELLER_PRODUCT_LIFECYCLE_SERVICE,
+          useClass: SellerProductLifecycleService,
+        },
+        {
+          provide: SELLER_PRODUCT_IMAGE_SERVICE,
+          useClass: SellerProductImageService,
+        },
+        {
+          provide: SELLER_PRODUCT_TAXONOMY_SERVICE,
+          useClass: SellerProductTaxonomyService,
+        },
         SellerOptionService,
         SellerCustomTemplateService,
         SellerRepository,
